% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_pt_beta.R
\name{background_pt}
\alias{background_pt}
\title{Generate Background Points for Habitat Suitability Modeling}
\usage{
background_pt(data = NULL, rast = NULL, n_bg = 0.1, width_cal_area = 1e+05)
}
\arguments{
\item{data}{An object created by \code{\link{prepare_occs}} with the cleaned occurrence data.}

\item{rast}{A raster object used for generating background points.}

\item{n_bg}{A numeric value indicating the proportion of background points to be generated relative to the non-NA cells in the raster.}

\item{width_cal_area}{A value indicating the buffer around occurrences for calculating the calibration area in km. (e.g., 100000 = 100 km)}
}
\value{
A list with the following components:
\describe{
\item{bg}{A data frame of generated background points.}
\item{pr_bg}{A data frame combining presence and background points.}
\item{points_pr_bg}{A \code{SpatialPointsDataFrame} object for the combined presence and background points.}
\item{area_M}{The calibration area as an \code{sf} object.}
\item{isna}{A data frame of points with NA values.}
\item{na_occs}{A data frame of occurrence points with NA values.}
\item{na_bg}{A data frame of background points with NA values.}
}
}
\description{
This function generates background points for habitat suitability modeling. Background points are taken from a calibration area (here defined by a convex hull with buffer)
(see, e.g., Hirzel & Le Lay, 2008). Additionally, this function does sampling bias correction using a so-called thickening approach (Barber et al., 2022)
by creating a density grid based on occurrences and selecting background points according to density weights from within the calibration area.
The calibration area is created using the \code{calib_area} function from the \code{flexsdm} package v1.3.3 (Velazco et al., 2022). A density grid is created using the \code{sp.kde} function from the \code{spatialEco} package (Evans et al. 2017).
}
\examples{
\dontrun{
# Example usage
result <- background_pt(data = occs, rast = clim[[1]], n_bg = 0.1, width_cal_area = 100000)
print(result)
}

}
\references{
Velazco, S. J. E., Rose, M. B., de Andrade, A. F. A., Minoli, I. & de Alvarenga, A. A. (2022). flexsdm: An r package for supporting a comprehensive and flexible species distribution modelling workflow. Methods in Ecology and Evolution, 13(8), 1661-1669. doi:10.1111/2041-210X.13874

Evans, J. S., Murphy, M. A., & Ram, K. (2021). spatialEco: Spatial Analysis and Modelling Utilities. R package version 1.3-1. Retrieved from http://archive.linux.duke.edu/cran/web/packages/spatialEco/spatialEco.pdf
}
