% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_numbs_beta.R
\name{calc_numbs}
\alias{calc_numbs}
\title{Calculate Presence, Absence Counts, and Weights}
\usage{
calc_numbs(train_test)
}
\arguments{
\item{train_test}{A list object returned by \code{\link{test_train_aoi}}, containing training
and complete dataset components.}
}
\value{
A list with two elements, each a list for the training set and the complete set.
Each sub-list includes the number of presences (\code{prNum}), the number of absences (\code{bgNum}),
and a vector of weights (\code{casewts}) for each observation. Weights help in adjusting for
the imbalance between presences and absences.
}
\description{
This function calculates the number of presences and absences, as well as weights
to handle imbalanced data in species distribution modeling or similar contexts.
Weights are adjusted to balance between presence and background (absence) cases.
}
\details{
The approach and parts of the code were adapted from Valavi et al., 2022.
}
\examples{
\dontrun{
# Assume `test_train_aoi` has been run previously to generate `train_test`
results <- calc_numbs(train_test=tta)
print(results)
}

}
\references{
Valavi, R., Guillera-Arroita, G., Lahoz-Monfort, J. J., & Elith, J. (2022).
Predictive performance of presence-only species distribution models: A benchmark study with reproducible code.
Ecological Monographs, 92(1), e1486. https://doi.org/10.1002/ecm.1486
}
