% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_predict_beta.R
\name{create_ensemble_model}
\alias{create_ensemble_model}
\title{Generate Ensemble Predictions}
\usage{
create_ensemble_model(combine, i = i, tta)
}
\arguments{
\item{combine}{A list containing model outputs and predictions.}

\item{i}{Index for the current group of predictors within the combine list.}

\item{tta}{Testing set object containing observed occurrences.}
}
\description{
This function combines predictions from multiple models to generate an ensemble prediction,
considering only models with significant predictive value (AUC > 0.5). It calculates combined
ROC, AUC, COR, and creates an ensemble raster prediction.
}
