% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation_risk_assessment_beta.R
\name{extrapolation_risk_assessment}
\alias{extrapolation_risk_assessment}
\title{Extrapolation Risk Assessment for Habitat suitability modelling}
\usage{
extrapolation_risk_assessment(pred, backgr, st, test_train)
}
\arguments{
\item{pred}{Raster with the environmental predictors used for modelling.}

\item{backgr}{An object returned by \code{\link{background_pt}} with information about the calibration area (\code{area_M}) to be used for masking.}

\item{st}{A spatial object containing the coordinate reference system of the projection area.}

\item{test_train}{A object returned by \code{\link{test_train_aoi}}.}
}
\value{
A raster layer with extrapolation risk scores.
}
\description{
This function assesses the extrapolation risk of predictions based on the known calibration area.
It sources an external function 'exdet' from https://gist.github.com/johnbaums to perform the risk calculation.
}
\examples{
\dontrun{
result <- extrapolation_risk_assessment(pred = pred, backgr = backgr, st = combined_stack)
plot(result)
}
}
