% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_predict_lasso_beta.R
\name{fit_and_predict_lasso}
\alias{fit_and_predict_lasso}
\title{Fit Lasso Model on Ecological Data}
\usage{
fit_and_predict_lasso(training_test, numbs, rast_ext)
}
\arguments{
\item{training_test}{A list containing elements for training and testing sets.}

\item{numbs}{A list containing numerical parameters such as weights for training and full data sets.}

\item{rast_ext}{A raster layer used for masking output predictions.}
}
\value{
A structured list comprising two major components:
\itemize{
\item \verb{$eval}: Contains the outcomes and evaluations based on the training dataset. It includes:
\itemize{
\item \code{test_model}: The model object returned from training.
\item \code{pred}: Predictions made by the model on the test dataset.
\item \code{roc}: The ROC curve object indicating the performance of the model at various threshold settings.
\item \code{AUC}: The area under the ROC curve, a scalar measure of model performance.
\item \code{COR}: The correlation coefficient between observed outcomes and model predictions, indicating prediction accuracy.
\item \code{raster_test}: Raster predictions based on the training model, masked by \code{rast_ext}.
}
\item \verb{$full}: Contains the outcomes based on the complete dataset (training and testing combined). It includes:
\itemize{
\item \code{full_model}: The model object returned from training on the complete dataset.
\item \code{raster_full}: Raster predictions based on the full model, also masked by \code{rast_ext}.
taining model evaluations, predictions, and raster outputs.
}
}
}
\description{
This function fits a lasso regression model to ecological data. It prepares the data by creating quadratic terms,
fits the model on training data, and evaluates it on test data. It also performs predictions on a raster object.
The code sources external scripts from valavi/myspatial and relies on code provided by Valavi et al. 2022.
}
\examples{
\dontrun{
# Assuming training_test, numbs, and rast_ext are predefined:
results <- fit_lasso(training_test, numbs, rast_ext)
print(results)
}

}
