% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampsize_eval_beta.R
\name{perform_env_completeness_analysis}
\alias{perform_env_completeness_analysis}
\title{Evaluate Sample Size for Environmental Representativeness}
\usage{
perform_env_completeness_analysis(data_path, base_path, cores = 2, iter = 500)
}
\arguments{
\item{data_path}{Character string indicating the path to the climatic, soil, and topographic data files.}

\item{base_path}{Character string indicating the path to save the output files.}

\item{cores}{Integer specifying the number of cores to leave aside for parallel processing, with a default of 2.}

\item{iter}{Integer specifying the number of iterations for stochastic processes, with a default of 500.}
}
\value{
Saves the environmental completeness analysis results and plots to the specified output path.
}
\description{
This function assesses how representative occurrences are for the environmental conditions of a habitat.
It calculates the overlap between the full environmental space defined by all occurrence points and the environmental
space based on a proportional share of occurrences. Niche overlap is calculated using the Schoener D index.
A sample might be considered representative if, for example, less than 90\% of the occurrences retrieve more than 90\%
of the full environmental space based on all occurrences. This function utilizes the \code{ecospat} package to calculate
overlap between niches.
}
\examples{
perform_env_completeness_analysis("output/predictions", "data/processed", "output/tables/habitat_completness", 2)
}
