% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables_pergroup_beta.R
\name{select_variables_pergroup}
\alias{select_variables_pergroup}
\title{Select Variables Per Group for Habitat Suitability Modeling}
\usage{
select_variables_pergroup(groups = NULL, data = NULL)
}
\arguments{
\item{groups}{A list of raster stacks representing different groups of predictors.
All raster objects must have the same CRS (Coordinate Reference System) and extent.}

\item{data}{An object returned by \code{\link{background_pt}}.}
}
\value{
A list where each element corresponds to a group with selected variables and their
analysis metrics, including results of biserial correlation and VIF values.
}
\description{
This function selects relevant environmental variables from a specified dataset for each group,
typically representing different ecological groups or conditions. It performs variable selection by
analyzing data using biserial correlation and checking for multicollinearity (VIF) utilizing the
SDMworkshop package by BlasBenito (https://github.com/BlasBenito/SDMworkshop).
}
\examples{
\dontrun{
my_groups <- handle_error(list(clim = clim, notclim = stack(topo, soil_new)))
selection <- select_variables_pergroup(groups = groups, data = backgr)
print(selection)
}

}
