% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_train_aoi_beta.R
\name{test_train_aoi}
\alias{test_train_aoi}
\title{Create Training and Testing Datasets with Specified Area of Interest}
\usage{
test_train_aoi(
  data = NULL,
  predictors = NULL,
  partition = 0.8,
  aoi_width = NULL
)
}
\arguments{
\item{data}{An object returned by \code{\link{background_pt}}, containing spatial points data.}

\item{predictors}{A RasterStack or RasterBrick object representing the environmental variables used as predictors.}

\item{partition}{A numeric value between 0 and 1 indicating the proportion of the dataset to be used for training.
The default is 0.8, meaning 80\% of the data is used for training and the remaining 20\% for testing.}

\item{aoi_width}{A numeric value representing the distance in meters to create a bounding box around presence
and background points for projection. This bounding box defines the Area of Interest (AOI).
For example, 100000 represents a distance of 100 kilometers.}
}
\value{
A list with the following elements:
\itemize{
\item \code{training}: a sublist containing both the spatial points and the extracted predictor variables for the training set, confined to the AOI.
\item \code{testing}: a sublist containing both the spatial points and the extracted predictor variables for the testing set, confined to the AOI.
\item \code{comp}: a combined dataset of training and testing sets, both confined within the AOI.
\item \code{aoi}: the cropped area of the raster layer adjusted to the extent of the bounding box defining the AOI.
}
}
\description{
This function divides the input data into training and testing sets based on a specified partition ratio.
It extracts predictor variables for each set and adjusts the raster layers to the defined area of interest (AOI),
which is calculated using a minimum bounding box based on the extent of provided data points. The AOI specifies
the geographic area over which predictions should be applied.
}
\examples{
\dontrun{
result <- test_train_aoi(data = my_data, predictors = my_predictors, partition = 0.8, aoi_width = 100000)
print(result)
}
\dontrun{
result <- test_train_aoi(data = my_data, predictors = my_predictors, partition = 0.8, mbox = my_box)
print(result)
}
}
