% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_soil_beta.R
\name{tukey_soil}
\alias{tukey_soil}
\title{Perform Tukey's HSD Test for selecting most significcant soil layers}
\usage{
tukey_soil(
  pattern = NULL,
  points = NULL,
  soil = NULL,
  growthform = c("small_annual", "annual", "shrub", "tree")
)
}
\arguments{
\item{pattern}{A string specifying the soil depth layers to include. If NULL, a default pattern is selected based on the growth form. "0.5|5.15|15.30" selects soil layers from 0-5 cm, 5-15cm and 15-30 cm. Soil layers should be stored with the information for depth in the following way 0.5 for 0-5cm, 15.30 for 15-30 cm etc.}

\item{points}{An object created by \code{\link{background_pt}}.}

\item{soil}{A raster stack of soil layers from https://soilgrids.org/.}

\item{growthform}{A string specifying the growth form. Options are "small_annual", "annual", "shrub", or "tree".}
}
\value{
A modified raster stack with non-significant soil layers removed.
}
\description{
This function performs Tukey's Honest Significant Difference (HSD) test on soil layers for a given set of points and growth form. It identifies and removes non-significant soil layers based on the specified growth form or a custom pattern.
This function works only for soil layers from https://soilgrids.org/
}
\details{
This function uses Tukey's HSD test to compare the means of different soil layers and identify those that are not significantly different from each other. The test is conducted at a confidence level of 99\% (conf.level = 0.99) and a significance level of 5\% (p.adj > 0.05) for adjusted p-values.
}
\examples{
\dontrun{
result <- tukey_soil(pattern = NULL, points = backgr, soil = soil, growthform = "small_annual")
print(result)
}

}
